+++ tool_sideview/sideview_view.py
@@ -93,7 +93,8 @@ class SideViewPlotWidget(pg.PlotWidget):
 
         self.channel_bottom_plot = pg.PlotDataItem(np.array([(0.0, np.nan)]), pen=pen)
 
-        pen = pg.mkPen(color=QColor(255, 192, 203), width=4)
+        pen = pg.mkPen(color=QColor(100, 100, 100), width=4)
+        self.culvert_lowest_plot = pg.PlotDataItem(np.array([(0.0, np.nan)]), pen=pen)
         self.culvert_bottom_plot = pg.PlotDataItem(np.array([(0.0, np.nan)]), pen=pen)
         self.culvert_upper_plot = pg.PlotDataItem(np.array([(0.0, np.nan)]), pen=pen)
         self.culvert_top_plot = pg.PlotDataItem(np.array([(0.0, np.nan)]), pen=pen)
@@ -123,26 +124,35 @@ class SideViewPlotWidget(pg.PlotWidget):
         )
 
         # Add some structure specific fills
-        self.orifice_opening_fill = pg.FillBetweenItem(
-            self.orifice_upper_plot, self.orifice_middle_plot, pg.mkBrush(208, 240, 192)
+
+        self.culvert_lower_fill = pg.FillBetweenItem(
+            self.culvert_bottom_plot, self.culvert_lowest_plot, pg.mkBrush(100, 100, 100)
         )
-        self.orifice_full_fill = pg.FillBetweenItem(
-            self.orifice_top_plot, self.orifice_bottom_plot, pg.mkBrush(0, 255, 0)
+        self.culvert_middle_fill = pg.FillBetweenItem(
+            self.culvert_upper_plot, self.culvert_bottom_plot, pg.mkBrush(230, 230, 230)
+        )
+        self.culvert_upper_fill = pg.FillBetweenItem(
+            self.culvert_top_plot, self.culvert_upper_plot, pg.mkBrush(100, 100, 100)
         )
 
-        self.culvert_opening_fill = pg.FillBetweenItem(
-            self.culvert_upper_plot, self.culvert_bottom_plot, pg.mkBrush(255, 192, 203)
+        self.orifice_lower_fill = pg.FillBetweenItem(
+            self.orifice_middle_plot, self.orifice_bottom_plot, pg.mkBrush(51, 160, 44)
+        )
+        self.orifice_middle_fill = pg.FillBetweenItem(
+            self.orifice_upper_plot, self.orifice_middle_plot, pg.mkBrush(165, 230, 161)
+        )
+        self.orifice_upper_fill = pg.FillBetweenItem(
+            self.orifice_top_plot, self.orifice_upper_plot, pg.mkBrush(51, 160, 44)
         )
 
-        self.culvert_full_fill = pg.FillBetweenItem(
-            self.culvert_top_plot, self.culvert_upper_plot, pg.mkBrush(255, 20, 147)
+        self.weir_lower_fill = pg.FillBetweenItem(
+            self.weir_middle_plot, self.weir_bottom_plot, pg.mkBrush(227, 26, 28)
         )
-
-        self.weir_opening_fill = pg.FillBetweenItem(
-            self.weir_upper_plot, self.weir_middle_plot, pg.mkBrush(250, 217, 213)
+        self.weir_middle_fill = pg.FillBetweenItem(
+            self.weir_upper_plot, self.weir_middle_plot, pg.mkBrush(255, 153, 155)
         )
-        self.weir_full_fill = pg.FillBetweenItem(
-            self.weir_top_plot, self.weir_bottom_plot, pg.mkBrush(255, 0, 0)
+        self.weir_upper_fill = pg.FillBetweenItem(
+            self.weir_top_plot, self.weir_upper_plot, pg.mkBrush(227, 26, 28)
         )
 
         self.sewer_top_fill = pg.FillBetweenItem(
@@ -157,6 +167,7 @@ class SideViewPlotWidget(pg.PlotWidget):
         self.addItem(self.sewer_top_plot)
         self.addItem(self.sewer_exchange_plot)
         self.addItem(self.channel_bottom_plot)
+        self.addItem(self.culvert_lowest_plot)
         self.addItem(self.culvert_bottom_plot)
         self.addItem(self.culvert_upper_plot)
         self.addItem(self.culvert_top_plot)
@@ -169,12 +180,15 @@ class SideViewPlotWidget(pg.PlotWidget):
         self.addItem(self.orifice_middle_plot)
         self.addItem(self.orifice_top_plot)
         self.addItem(self.node_indicator_intersection_plot)
-        self.addItem(self.orifice_full_fill)
-        self.addItem(self.orifice_opening_fill)
-        self.addItem(self.weir_full_fill)
-        self.addItem(self.weir_opening_fill)
-        self.addItem(self.culvert_full_fill)
-        self.addItem(self.culvert_opening_fill)
+        self.addItem(self.orifice_upper_fill)
+        self.addItem(self.orifice_middle_fill)
+        self.addItem(self.orifice_lower_fill)
+        self.addItem(self.weir_upper_fill)
+        self.addItem(self.weir_middle_fill)
+        self.addItem(self.weir_lower_fill)
+        self.addItem(self.culvert_upper_fill)
+        self.addItem(self.culvert_middle_fill)
+        self.addItem(self.culvert_lower_fill)
         self.addItem(self.exchange_plot)
 
         # Set the z-order of the curves (note that fill take minimum of its two defining curve as z-value)
@@ -184,6 +198,7 @@ class SideViewPlotWidget(pg.PlotWidget):
         self.sewer_top_plot.setZValue(10)
         self.sewer_exchange_plot.setZValue(10)
         self.channel_bottom_plot.setZValue(10)
+        self.culvert_lowest_plot.setZValue(10)
         self.culvert_bottom_plot.setZValue(10)
         self.culvert_upper_plot.setZValue(10)
         self.culvert_top_plot.setZValue(10)
@@ -198,13 +213,16 @@ class SideViewPlotWidget(pg.PlotWidget):
 
         self.exchange_plot.setZValue(100)
         self.node_indicator_intersection_plot.setZValue(55)
-        self.orifice_full_fill.setZValue(20)
-        self.orifice_opening_fill.setZValue(21)
-        self.weir_full_fill.setZValue(20)
-        self.weir_opening_fill.setZValue(21)
-        self.culvert_full_fill.setZValue(20)
-        self.culvert_opening_fill.setZValue(21)
-        self.bottom_fill.setZValue(3)
+        self.orifice_upper_fill.setZValue(20)
+        self.orifice_middle_fill.setZValue(3)
+        self.orifice_lower_fill.setZValue(20)
+        self.weir_upper_fill.setZValue(20)
+        self.weir_middle_fill.setZValue(3)
+        self.weir_lower_fill.setZValue(20)
+        self.culvert_upper_fill.setZValue(20)
+        self.culvert_middle_fill.setZValue(3)
+        self.culvert_lower_fill.setZValue(20)
+        self.bottom_fill.setZValue(7)
         self.sewer_top_fill.setZValue(3)
 
         # set listeners to signals
@@ -240,6 +258,7 @@ class SideViewPlotWidget(pg.PlotWidget):
         middle_line = []  # Typically crest-level
         exchange_line = []  # exchange level
         upper_limit_line = []  # For top fill of weirs, orifices and culverts
+        lower_limit_line = []  # For bottom of culverts
 
         self.current_grid_id = current_grid.id if current_grid else None
 
@@ -290,6 +309,8 @@ class SideViewPlotWidget(pg.PlotWidget):
                         upper_line.append((end_dist, end_level + end_height, ltype))
                         upper_limit_line.append((begin_dist, UPPER_LIMIT, ltype))
                         upper_limit_line.append((end_dist, UPPER_LIMIT, ltype))
+                        lower_limit_line.append((begin_dist, LOWER_LIMIT, ltype))
+                        lower_limit_line.append((end_dist, LOWER_LIMIT, ltype))
                     else:
                         upper_line.append((begin_dist, begin_level + begin_height, ltype))
                         upper_line.append((end_dist, end_level + end_height, ltype))
@@ -419,11 +440,22 @@ class SideViewPlotWidget(pg.PlotWidget):
             for point in upper_limit_line:
                 tables[point[2]].append((point[0], point[1]))
 
-            logger.info("---------")
-            logger.info(tables[LineType.CULVERT])
             self.culvert_top_plot.setData(np.array(tables[LineType.CULVERT], dtype=float), connect="pairs")
             self.weir_top_plot.setData(np.array(tables[LineType.WEIR], dtype=float), connect="pairs")
             self.orifice_top_plot.setData(np.array(tables[LineType.ORIFICE], dtype=float), connect="pairs")
+
+            tables = {
+                LineType.PIPE: [],
+                LineType.CHANNEL: [],
+                LineType.CULVERT: [],
+                LineType.PUMP: [],
+                LineType.WEIR: [],
+                LineType.ORIFICE: [],
+            }
+
+            for point in lower_limit_line:
+                tables[point[2]].append((point[0], point[1]))
+            self.culvert_lowest_plot.setData(np.array(tables[LineType.CULVERT], dtype=float), connect="pairs")
 
             # draw nodes
             node_indicator_table = []
@@ -475,6 +507,8 @@ class SideViewPlotWidget(pg.PlotWidget):
             self.channel_bottom_plot.setData(ts_table)
             self.culvert_bottom_plot.setData(ts_table)
             self.culvert_upper_plot.setData(ts_table)
+            self.culvert_lowest_plot.setData(ts_table)
+            self.culvert_top_plot.setData(ts_table)
             self.weir_bottom_plot.setData(ts_table)
             self.weir_upper_plot.setData(ts_table)
             self.weir_middle_plot.setData(ts_table)
@@ -519,7 +553,7 @@ class SideViewPlotWidget(pg.PlotWidget):
             water_level_plot = pg.PlotDataItem(np.array([(0.0, np.nan)]), pen=pen)
             water_level_plot.setZValue(100)  # always visible
             water_fill = pg.FillBetweenItem(water_level_plot, self.absolute_bottom, pg.mkBrush(plot_color[0], plot_color[1], plot_color[2], 128))
-            water_fill.setZValue(0)
+            water_fill.setZValue(5)
             water_level_nodes = pg.PlotDataItem(np.array([(0.0, np.nan)]), symbolBrush=pg.mkBrush(plot_color[0], plot_color[1], plot_color[2]), symbolSize=7)
             water_level_nodes.setZValue(100)
             self.addItem(water_level_plot)